Implementation of rent and leave done, returned response changed from custom to ResponseEntity, to get rid of bug in postman.

+++ src/main/java/com/bike/rental/controller/UserController.java
@@ -1,9 +1,12 @@
 package com.bike.rental.controller;
-import com.bike.rental.model.ApiResponse;
+
 import com.bike.rental.model.User;
 import com.bike.rental.service.UserService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -12,37 +15,54 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/users")
 public class UserController {
+    public static final Logger logger = LoggerFactory.getLogger(UserController.class);
 
     @Autowired
     private UserService userService;
 
-    @RequestMapping(value="/register", method = RequestMethod.POST)
-    public ApiResponse<User> saveUser(@RequestBody User user){
+    @RequestMapping(value = "/register", method = RequestMethod.POST)
+    public ResponseEntity<User> saveUser(@RequestBody User user) {
         User userObj = userService.findByEmail(user.getEmail());
-        if(userObj != null){
-            return new ApiResponse<>(HttpStatus.CONFLICT.value(), String.format("User %s is already registered.",user.getEmail()) , userObj);
+        if (userObj != null) {
+            logger.warn("User {} is already registered!", user.getEmail());
+            return new ResponseEntity<>(userObj, HttpStatus.CONFLICT);
         }
+
         user.setRentedBikeId(0L);
         userObj = userService.save(user);
-        return new ApiResponse<>(HttpStatus.OK.value(), "User is fetched successfully.",userObj);
+        logger.info("User {} successfully registered.", user.getEmail());
+        return new ResponseEntity<>(userObj, HttpStatus.OK);
     }
 
     @GetMapping
-    public ApiResponse<List<User>> getUsers(){
-        return new ApiResponse<>(HttpStatus.OK.value(), "User list fetched successfully.",userService.findAll());
+    public ResponseEntity<List<User>> getUsers() {
+        List<User> userList = userService.findAll();
+
+        if (userList == null || userList.isEmpty()) {
+            return new ResponseEntity<>(userList, HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(userList, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ApiResponse<User> getOne(@PathVariable Long id){
-        return new ApiResponse<>(HttpStatus.OK.value(), "User fetched successfully.",userService.findById(id));
+    public ResponseEntity<User> getOne(@PathVariable Long id) {
+        User user = userService.findById(id);
+        if (user == null) {
+            return new ResponseEntity<>(user, HttpStatus.NO_CONTENT);
+        }
+        return new ResponseEntity<>(user, HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ApiResponse<Void> delete(@PathVariable Long id) {
-        userService.delete(id);
-        return new ApiResponse<>(HttpStatus.OK.value(), "User deleted successfully.", null);
-    }
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        User user = userService.findById(id);
+        if (user == null) {
+            return new ResponseEntity<>(String.format("User with id {} doesn't exist!", id), HttpStatus.BAD_REQUEST);
+        }
 
+        userService.delete(id);
+        return new ResponseEntity<>(String.format("User with id {} removed successfully.", id), HttpStatus.OK);
+    }
 
 
 }
